requirements.txt
psycopg2-binary
streamlit
google-generativeai
python-dotenv
sqlalchemy